% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-method.R
\name{coef.tsci}
\alias{coef.tsci}
\title{Extract Model Coefficients of TSCI Fits}
\usage{
\method{coef}{tsci}(object, all = FALSE, ...)
}
\arguments{
\item{object}{a object of class 'tsci'.}

\item{all}{logical. If \code{FALSE}, only the treatment effect estimate of the
selected violation space candidate(s) is returned. If \code{TRUE}, additionally the treatment
effect estimates of each violation space candidate are returned.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Coefficients extracted form the model object \code{object}.
}
\description{
Extract Model Coefficients of TSCI Fits
}
