% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monomials.R
\name{create_monomials}
\alias{create_monomials}
\title{Monomials as Violation Space Candidates}
\usage{
create_monomials(Z, degree, type)
}
\arguments{
\item{Z}{observations of the instrumental variable(s). Either a numeric vector of length n
or a numeric matrix with dimension n by s.}

\item{degree}{The degree up to which monomials should be created.}

\item{type}{One out of \code{monomials_main} or \code{monomials_full}. \cr
\code{monomials_main} creates the monomials for the polynomials of each instrumental variable up to degree \code{degree}. \cr
\code{monomials_full} creates the monomials for the polynomials of a combination of all instrumental variables up to degree \code{degree}.}
}
\value{
A ordered list. Each element is a matrix consisting of the monomials to be added to
the next violation space candidate.
}
\description{
Monomials as Violation Space Candidates
}
\details{
assuming there are 3 instrumental variables Z1, Z2, and Z3 and \code{degree} = d,
\code{monomials_main} creates the monomials of the polynomials (Z1 + 1)^d, (Z2 + 1)^d, (Z3 + 1)^d without the constants and
\code{monomials_full} creates the monomials of the polynomials (Z1 + Z2 + Z3)^1, (Z1 + Z2 + Z3)^2, ..., (Z1 + Z2 + Z3)^d without the constants.
Thus \code{type} = \code{monomials_main} does not include interactions between the instrumental variables.
}
\examples{
Z <- matrix(rnorm(100 * 3), nrow = 100, ncol = 3)
vio_space <- create_monomials(Z = Z, degree = 4, type = "monomials_full")
}
