% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_oob.R
\name{forest_oob}
\alias{forest_oob}
\title{hyperparameter tuning using out of bag mean squared error.}
\usage{
forest_oob(Data_A2, params_grid)
}
\arguments{
\item{Data_A2}{a data frame containing the subset of observations used for fitting the treatment model.}

\item{params_grid}{a data frame containing the combination of hyperparameter values.}
}
\value{
a list containing the hyperparameter values chosen by cv and the cv mse.
}
\description{
hyperparameter tuning using out of bag mean squared error.
}
\examples{
params_grid <- expand.grid(num_trees = 200,
mtry = 0.5,
max_depth = 0,
min_node_size = c(1, 10))
Z <- rnorm(100)
X <- rnorm(100)
D <- 2 * Z + X + rnorm(100)
Data <- data.frame("D" = D, "Z" = Z, "X" = X)
cv_fit <- forest_oob(Data_A2 = Data, params_grid = params_grid)
cv_fit$params_A2
}
