% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_hatmatrix.R
\name{get_forest_hatmatrix}
\alias{get_forest_hatmatrix}
\title{Title}
\usage{
get_forest_hatmatrix(Data_A1, Data_A2, params)
}
\arguments{
\item{Data_A1}{a data frame containing the subset of observations used for fitting the outcome model.}

\item{Data_A2}{a data frame containing the subset of observations used for fitting the treatment model.}

\item{params}{a list containing the hyperparameters of the boosting method.}
}
\value{
a list containing the following the following components: \tabular{ll}{
\code{weight} \tab the hat matrix. \cr
\tab \cr
\code{forest_A2} \tab the fitted random forest model. \cr
}
}
\description{
Title
}
\examples{
Z <- rnorm(100)
X <- rnorm(100)
D <- 2 * Z + X + rnorm(100)
Data <- data.frame("D" = D, "Z" = Z, "X" = X)
Data_A1 <- Data[c(1:50), ]
Data_A2 <- Data[-c(1:50), ]
params_forest <- list("num_trees" = 200, "mtry" = 1, "max_depth" = 0, "min_node_size" = 5)
hat_matrix_forest <- get_forest_hatmatrix(Data_A1 = Data_A1, Data_A2 = Data_A2,
  params = params_forest)
mean((Data_A1$D - hat_matrix_forest$weight \%*\% Data_A1$D)^2)
}
