% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hat_matrix_of_tree.R
\name{get_hat_matrix_of_tree}
\alias{get_hat_matrix_of_tree}
\title{Calculates the hat matrix of a tree.}
\usage{
get_hat_matrix_of_tree(leaves)
}
\arguments{
\item{leaves}{A character or continuous vector specifying the leaf membership.
Entries with the same values are assumed to belong in the same leaf.}
}
\value{
A n x n continuous vector. Hat matrix of the tree.
}
\description{
Calculates the hat matrix of a tree.
}
\examples{
leaves <- c("A", "B", "B", "C", "B", "B", "A")
get_hat_matrix_of_tree(leaves)
}
