% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l2boost_hat_matrix.R
\name{get_l2boost_hat_matrix}
\alias{get_l2boost_hat_matrix}
\title{Calculates the hat matrix of the boosting fit.}
\usage{
get_l2boost_hat_matrix(Data_A1, Data_A2, params, xgboost)
}
\arguments{
\item{Data_A1}{A data frame containing the subset of observations used for fitting the outcome model.}

\item{Data_A2}{A data frame containing the subset of observations used for fitting the treatment model.}

\item{params}{A list containing the hyperparameters of the boosting method.}

\item{xgboost}{logical, specifies whether xgboost should be used to fit the model.}
}
\value{
A list containing the hat matrix and the fitted treatment model.
}
\description{
Calculates the hat matrix of the boosting fit.
}
\examples{
Z <- rnorm(100)
X <- rnorm(100)
D <- 2 * Z + X + rnorm(100)
Data <- data.frame("D" = D, "Z" = Z, "X" = X)
Data_A1 <- Data[c(1:50), ]
Data_A2 <- Data[-c(1:50), ]
params_xgboost <- list("nrounds" = 10, "eta" = 0.3, "max_depth" = 2, "subsample" = 1,
"colsample_bytree" = 1)
hat_matrix_xgboost <- get_l2boost_hat_matrix(Data_A1 = Data_A1, Data_A2 = Data_A2,
params = params_xgboost, xgboost = TRUE)
params_vanilla <- list("nrounds" = 10, "eta" = 0.3, "max_depth" = 2, "subsample" = 1,
"colsample_bytree" = 1)
hat_matrix_vanilla <- get_l2boost_hat_matrix(Data_A1 = Data_A1, Data_A2 = Data_A2,
params = params_vanilla, xgboost = FALSE)
mean((Data_A1$D - hat_matrix_xgboost$weight \%*\% Data_A1$D)^2)
mean((Data_A1$D - hat_matrix_vanilla$weight \%*\% Data_A1$D)^2)
}
