% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_l2boost_hatmatrix.R
\name{get_l2boost_hatmatrix}
\alias{get_l2boost_hatmatrix}
\title{Calculates the hat matrix of the boosting fit.}
\usage{
get_l2boost_hatmatrix(Data_A1, Data_A2, params, xgboost)
}
\arguments{
\item{Data_A1}{a data frame containing the subset of observations used for fitting the outcome model.}

\item{Data_A2}{a data frame containing the subset of observations used for fitting the treatment model.}

\item{params}{a list containing the hyperparameters of the boosting method.}

\item{xgboost}{logical, specifies whether xgboost should be used to fit the model.}
}
\value{
a list containing the following the following components: \tabular{ll}{
\code{weight} \tab the hat matrix. \cr
\tab \cr
\code{treeboost_A2} \tab the fitted boosting model. \cr
}
}
\description{
Calculates the hat matrix of the boosting fit.
}
\examples{
Z <- rnorm(100)
X <- rnorm(100)
D <- 2 * Z + X + rnorm(100)
Data <- data.frame("D" = D, "Z" = Z, "X" = X)
Data_A1 <- Data[c(1:50), ]
Data_A2 <- Data[-c(1:50), ]
params_xgboost <- list("nrounds" = 10, "eta" = 0.3, "max_depth" = 2, "subsample" = 1,
  "colsample_bytree" = 1)
hat_matrix_xgboost <- get_l2boost_hatmatrix(Data_A1 = Data_A1, Data_A2 = Data_A2,
  params = params_xgboost, xgboost = TRUE)
params_vanilla <- list("nrounds" = 10, "eta" = 0.3, "max_depth" = 2, "subsample" = 1,
  "colsample_bytree" = 1)
hat_matrix_vanilla <- get_l2boost_hatmatrix(Data_A1 = Data_A1, Data_A2 = Data_A2,
  params = params_vanilla, xgboost = FALSE)
mean((Data_A1$D - hat_matrix_xgboost$weight \%*\% Data_A1$D)^2)
mean((Data_A1$D - hat_matrix_vanilla$weight \%*\% Data_A1$D)^2)
}
