% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2boost_cv.R
\name{l2boost_cv}
\alias{l2boost_cv}
\title{Performs hyperparameter selection by cross-validation.}
\usage{
l2boost_cv(Data_A2, params_grid, nfolds, xgboost)
}
\arguments{
\item{Data_A2}{a data frame containing the subset of observations used for fitting the treatment model.}

\item{params_grid}{a data frame containing the combination of hyperparameter values.}

\item{nfolds}{numeric, the number of folds for the cross validation.}

\item{xgboost}{logical, specifies whether xgboost should be used to fit the model.}
}
\value{
a list containing the hyperparameter values chosen by cv and the cv mse.
}
\description{
Performs hyperparameter selection by cross-validation.
}
\examples{
params_grid <- expand.grid(nrounds = 50,
eta = 0.3,
max_depth = c(2, 6),
subsample = c(0.5, 1),
colsample_bytree = 1,
early_stopping = TRUE)
Z <- rnorm(100)
X <- rnorm(100)
D <- 2 * Z + X + rnorm(100)
Data <- data.frame("D" = D, "Z" = Z, "X" = X)
cv_fit <- l2boost_cv(Data_A2 = Data, params_grid = params_grid, nfolds = 10, xgboost = TRUE)
cv_fit$params_A2
}
