% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary-method.R
\name{print.summary.tsci}
\alias{print.summary.tsci}
\title{Title}
\usage{
\method{print}{summary.tsci}(object, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
Title
}
