% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-method.R
\name{print.tsci}
\alias{print.tsci}
\title{Print Content of tsci Object.}
\usage{
\method{print}{tsci}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class 'tsci'.}

\item{digits}{number of significant digits to display.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Print Content of tsci Object.
}
