% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basis-Approach.R
\name{TSCI.basis.fit}
\alias{TSCI.basis.fit}
\title{Title}
\usage{
TSCI.basis.fit(D, W, folds = 5)
}
\arguments{
\item{D}{continuous or binary, n by 1 treatment vector.}

\item{W}{continuous or binary, instrument-covariates matrix.}

\item{folds}{number of folds for cross validation,default by 5.}
}
\value{
\item{\code{D.rep}}{representation of D.}
    \item{\code{SigmaSqD}}{estimated noise level in treatment model.}
    \item{\code{knot}}{number of knots.}
    \item{\code{M}}{number of basis functions, equals to #knot + 2.}
}
\description{
Title
}
\examples{
\donttest{

}
}
