% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random-Forest.R
\name{TSCI.RF.fit}
\alias{TSCI.RF.fit}
\title{Random Forest with data splitting}
\usage{
TSCI.RF.fit(
  D,
  Z,
  X,
  num.trees,
  mtry,
  max.depth,
  min.node.size,
  split.prop,
  MSE.thol = 1e+12,
  forest.save = T
)
}
\arguments{
\item{D}{continuous or binary, n by 1 treatment vector.}

\item{Z}{continuous or binary, n by 1 Instrumental Variable, only one instrument is implemented for violation space selection now.}

\item{X}{continuous or binary, n by p_x baseline covariates matrix.}

\item{num.trees}{number of trees.}

\item{mtry}{number of covariates to possibly split at in each node.}

\item{max.depth}{maximal depth of each tree.}

\item{min.node.size}{minimal size of each leaf node.}

\item{split.prop}{a value between 0 and 1, the proportion of samples we use in A1.}

\item{MSE.thol}{a very large value of MSE, default by 1e12, used for the start of hyper-parameter selection.}

\item{forest.save}{save the Random Forest output or not, default by TRUE.}
}
\value{
\item{\code{forest.A2}}{random forest built on subsample A2, available if forest.save=TRUE.}
    \item{\code{params.A2}}{best hyper-parameters for forest.A2 selected by out-of-bag error.}
    \item{\code{A1.ind}}{indices of subsample A1.}
    \item{\code{nodes.A1}}{a n_A1 by num.trees matrix, rows refer to different samples, columns refer to different trees, the entrees are leaf node indices of each sample in each tree.}
    \item{\code{MSE.oob}}{minimal out-of-bag error using the best hyper-parameters.}
}
\description{
Random Forest with data splitting
}
\examples{
\donttest{

}


}
