% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random-Forest.R
\name{TSCI.RF.weight}
\alias{TSCI.RF.weight}
\title{Weight matrix computation of random forest}
\usage{
TSCI.RF.weight(nodes)
}
\arguments{
\item{nodes}{A n_A1 by num.trees matrix, rows refer to different samples, columns refer to different trees, the entrees are leaf node indices of each sample in each tree.}
}
\value{
\item{\code{out.weight}}{A n_A1 by n_A1 symmetric sparse weight matrix of class dgCMatrix with the ith row represents the weights of outcome of each sample on the prediction of the ith outcome.}
}
\description{
Weight matrix computation of random forest
}
\examples{
\donttest{

}
}
