% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsci_forest.R
\name{tsci_forest}
\alias{tsci_forest}
\title{Two stage curvature identification with random forest}
\usage{
tsci_forest(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  vio_space = NULL,
  layer = TRUE,
  split_prop = 2/3,
  num_trees = NULL,
  mtry = NULL,
  max_depth = NULL,
  min_node_size = NULL,
  forest_save = TRUE,
  str_thol = 10,
  alpha = 0.05,
  alternative_bias_adjustment = FALSE,
  multi_splitting = FALSE,
  nsplits = NULL,
  mult_split_method = "Meinshausen"
)
}
\arguments{
\item{Y}{the outcome vector.}

\item{D}{the treatment vector.}

\item{Z}{the instrumental variable vector.}

\item{X}{the baseline covariates.}

\item{intercept}{logical, if \code{TRUE} an intercept is included in the outcome model.}

\item{vio_space}{list or matrix containing the violation space.}

\item{layer}{logical, if \code{TRUE} violation space selection is performed.}

\item{split_prop}{numeric, proportion of observations used to fit the outcome model.}

\item{num_trees}{numeric, hyperparameter of the random forest algorithm. Specifies the number of trees.}

\item{mtry}{numeric, hyperparameter of the random forest algorithm. Specifies the number of variables to possibly split at in each node.}

\item{max_depth}{numeric, hyperparameter of the random forest algorithm. Specifies the maximum tree depth. A value of NULL or 0 (the default) corresponds to unlimited depth.}

\item{min_node_size}{numeric, hyperparameter of the random forest algorithm. Specifies the minimal node size.}

\item{forest_save}{logical, specifies if the fitted random forest model should be returned.}

\item{str_thol}{numeric, the minimal value of the threshold of IV strength test.}

\item{alpha}{numeric, specifies the desired significance level.}

\item{alternative_bias_adjustment}{logical, if \code{TRUE} a slightly
different bias adjustment is used.}

\item{multi_splitting}{logical, if \code{TRUE} multi-splitting will be performed.}

\item{nsplits}{numeric, number of times the data will be split.}

\item{mult_split_method}{method to for inference if multi-splitting is performed. Either 'Chernozhukov' or 'Meinshausen'.}
}
\value{
a list containing at least the following the following components: \tabular{ll}{
\code{Coef_vec} \tab a named vector containing the coefficients for all tested violation spaces. \cr
\tab \cr
\code{sd_vec} \tab a named vector containing the standard errors of the coefficients
for all tested violation spaces. \cr
\tab \cr
\code{Coef_robust} \tab a named vector containing the coefficients of the selected violation space. \cr
\tab \cr
\code{sd_robust} \tab a named vector containing the standard errors of the coefficients
of the selected violation space. \cr
\tab \cr
\code{CI_robust} \tab a named vector containing the confidence intervals of the coefficients
of the selected violation space. \cr
\tab \cr
\code{iv_str} \tab a named vector containing the estimated instrumental variable strengths
of all tested violation spaces. \cr
\tab \cr
\code{iv_str} \tab a named vector containing the estimated necessary instrumental variable strengths
of all tested violation spaces. \cr
\code{SigmaSqD} \tab the estimated variance of the treatment model. \cr
\tab \cr
\code{SigmaSqY} \tab a named vector containing the estimated variances of the treatment model
of all tested violation spaces. \cr
\tab \cr
\code{SigmaSqY.Qmax} \tab the estimated variance of the treatment model
of the largest considered violation space. \cr
\tab \cr
\code{trace_T} \tab the trace of the M matrix. \cr
\tab \cr
\code{explained_iv} \tab a named vector containing the numerators of the estimated standard errors of the coefficients
of the selected violation space. \cr
\tab \cr
\code{Qmax} \tab the number of the largest considered violation space \cr
\tab \cr
\code{q_comp} \tab the number of the selected violation space \cr
\tab \cr
\code{q_robust} \tab the number of the selected violation space by the robust approach. \cr
\tab \cr
\code{invalidity} \tab logical, if \code{TRUE} the hypothesis of a valid instrument is rejected. \cr
\tab \cr
\code{run_OLS} \tab logical, if \code{TRUE} the instrument is weak even if it's assumed to be valid. \cr
\tab \cr
\code{run_OLS} \tab logical, if \code{TRUE} the instrument is weak for all tested non empty violation spaces. \cr
\tab \cr
\code{mse_oob} \tab the out-of-bag mean squared error of the first stage.
}
}
\description{
Two stage curvature identification with random forest
}
\examples{
n <- 100
Z <- rnorm(n)
X <- rnorm(n)
D <- Z + 2 * Z^2 + X + rnorm(n)
Y <- D + Z + X + rnorm(n)
tsci_forest(Y = Y, D = D, Z = Z, X = X)
}
