% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsci_second_stage_stats.R
\name{tsci_second_stage_stats}
\alias{tsci_second_stage_stats}
\title{Computes the necessary statistics for TSCI.}
\usage{
tsci_second_stage_stats(
  D_rep,
  Cov_rep,
  weight,
  n,
  eps_hat,
  delta_hat,
  str_thol
)
}
\arguments{
\item{D_rep}{a vector denoting the treatment representation.}

\item{Cov_rep}{a matrix denoting the augmented instruments-covariates representation.}

\item{weight}{a matrix, the hat matrix obtained in the first stage.}

\item{n}{numeric, the full sample size.}

\item{eps_hat}{a vector containing the estimated noise of the outcome model.}

\item{delta_hat}{a vector containing the estimated noise of the treatment model.}

\item{str_thol}{numeric, the minimal value of the threshold of IV strength test.}
}
\value{
a list containing the following the following components: \tabular{ll}{
\code{sd} \tab the estimated standard error of the treatment parameter estimate. \cr
\tab \cr
\code{D_resid} \tab BLANK. \cr
\tab \cr
\code{iv_str} \tab the estimated instrumental variable strength. \cr
\tab \cr
\code{iv_thol} \tab the estimated necessary instrumental variable strength. \cr
\tab \cr
\code{explained_iv} \tab the numerator of the estimated standard error. \cr
\tab \cr
\code{trace_T} \tab the trace of the M matrix. \cr
\tab \cr
\code{RSS_V} \tab the diagonal entries of the M matrix. \cr
}
}
\description{
Computes the necessary statistics for TSCI.
}
\examples{
NULL
}
