% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-method.R
\name{coef.tsci}
\alias{coef.tsci}
\title{Extract Model Coefficients of TSCI Fits.}
\usage{
\method{coef}{tsci}(object, parm = NULL, ...)
}
\arguments{
\item{object}{an object of class 'tsci'.}

\item{parm}{a specification for which treatment effect estimates should be returned.
Either a vector of numbers or a vector of names or 'all'.
If missing, the treatment effect estimate by violation space selection is returned.
If 'all', the treatment effect estimates for all violation space candidates are returned.}

\item{...}{arguments to be passed to or from other methods.}
}
\value{
Coefficients extracted form the model object \code{object}.
}
\description{
Extract Model Coefficients of TSCI Fits.
}
