% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-method.R
\name{confint.tsci}
\alias{confint.tsci}
\title{Confidence Intervals of Treatment Effect Estimates for TSCI Fits.}
\usage{
\method{confint}{tsci}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a object of class 'tsci'.}

\item{parm}{a specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names or 'all'.
If missing, the confidence interval for the selected violation space candidate of the
robust method is returned. If 'all', the confidence intervals for all violation space candidates
are returned.}

\item{level}{the confidence level required.}

\item{...}{additional argument(s) for methods.}
}
\value{
a matrix containing the confidence intervals.
}
\description{
Confidence Intervals of Treatment Effect Estimates for TSCI Fits.
}
