% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint-method.R
\name{confint.tsci}
\alias{confint.tsci}
\title{xxx}
\usage{
\method{confint}{tsci}(object, level, ...)
}
\arguments{
\item{object}{xxx}

\item{level}{xxx}

\item{...}{xxx}
}
\value{
xxx
}
\description{
xxx
}
