% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-summary-method.R
\name{print.summary.tsci}
\alias{print.summary.tsci}
\title{Print Content of summary.tsci Object.}
\usage{
\method{print}{summary.tsci}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class 'summary.tsci'.}

\item{digits}{number of significant digits to display.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Print Content of summary.tsci Object.
}
