% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsci_boosting.R
\name{tsci_boosting}
\alias{tsci_boosting}
\title{Calculates the hat matrix of the first stage.}
\usage{
tsci_boosting(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  vio_space = NULL,
  layer = TRUE,
  split_prop = 2/3,
  nrounds = NULL,
  eta = NULL,
  max_depth = NULL,
  subsample = NULL,
  colsample_bytree = NULL,
  early_stopping = NULL,
  nfolds = 10,
  l2boost_save = TRUE,
  str_thol = 20,
  alpha = 0.05,
  xgboost = TRUE,
  alternative_bias_adjustment = FALSE,
  multi_splitting = FALSE,
  nsplits = NULL,
  mult_split_method = "Meinshausen"
)
}
\arguments{
\item{Y}{the outcome vector.}

\item{D}{the treatment vector.}

\item{Z}{the instrumental variable vector.}

\item{X}{the baseline covariates.}

\item{intercept}{logical, if \code{TRUE} an intercept is included in the outcome model.}

\item{vio_space}{list or matrix containing the violation space.}

\item{layer}{logical, if \code{TRUE} violation space selection is performed.}

\item{split_prop}{numeric, proportion of observations used to fit the outcome model.}

\item{nrounds}{numeric, hyperparameter of the boosting algorithm. Specifies the number of boosting iterations.}

\item{eta}{numeric, hyperparameter of the boosting algorithm. Specifies the learning rate.}

\item{max_depth}{numeric, hyperparameter of the boosting algorithm. Specifies the maximal depth of each tree.}

\item{subsample}{numeric, hyperparameter of the boosting algorithm. Specifies proportion of observations used to fit each tree.}

\item{colsample_bytree}{numeric, hyperparameter of the boosting algorithm. Specifies proportion of variables used to fit each tree.}

\item{early_stopping}{logical, hyperparameter of the boosting algorithm. If \code{TRUE} early stopping will be applied.}

\item{nfolds}{numeric, the number of folds for the k-fold cross validation.}

\item{l2boost_save}{logical, specifies if the fitted boosting model should be returned. Currently not implemented.}

\item{str_thol}{numeric, the minimal value of the threshold of IV strength test.}

\item{alpha}{numeric, specifies the desired significance level.}

\item{xgboost}{logical, specifies whether xgboost should be used to calculate the hat matrix.}

\item{alternative_bias_adjustment}{logical, if \code{TRUE} a slightly
different bias adjustment is used.}

\item{multi_splitting}{logical, if \code{TRUE} multi-splitting will be performed.}

\item{nsplits}{numeric, number of times the data will be split.}

\item{mult_split_method}{method to for inference if multi-splitting is performed. Either 'Chernozhukov' or 'Meinshausen'.}
}
\value{
a list containing at least the following the following components: \tabular{ll}{
\code{Coef_vec} \tab a named vector containing the coefficients for all tested violation spaces. \cr
\tab \cr
\code{sd_vec} \tab a named vector containing the standard errors of the coefficients
for all tested violation spaces. \cr
\tab \cr
\code{Coef_robust} \tab a named vector containing the coefficients of the selected violation space. \cr
\tab \cr
\code{sd_robust} \tab a named vector containing the standard errors of the coefficients
of the selected violation space. \cr
\tab \cr
\code{CI_robust} \tab a named vector containing the confidence intervals of the coefficients
of the selected violation space. \cr
\tab \cr
\code{iv_str} \tab a named vector containing the estimated instrumental variable strengths
of all tested violation spaces. \cr
\tab \cr
\code{iv_str} \tab a named vector containing the estimated necessary instrumental variable strengths
of all tested violation spaces. \cr
\code{SigmaSqD} \tab the estimated variance of the treatment model. \cr
\tab \cr
\code{SigmaSqY} \tab a named vector containing the estimated variances of the treatment model
of all tested violation spaces. \cr
\tab \cr
\code{SigmaSqY.Qmax} \tab the estimated variance of the treatment model
of the largest considered violation space. \cr
\tab \cr
\code{trace_T} \tab the trace of the M matrix. \cr
\tab \cr
\code{explained_iv} \tab a named vector containing the numerators of the estimated standard errors of the coefficients
of the selected violation space. \cr
\tab \cr
\code{Qmax} \tab the number of the largest considered violation space \cr
\tab \cr
\code{q_comp} \tab the number of the selected violation space \cr
\tab \cr
\code{q_robust} \tab the number of the selected violation space by the robust approach. \cr
\tab \cr
\code{invalidity} \tab logical, if \code{TRUE} the hypothesis of a valid instrument is rejected. \cr
\tab \cr
\code{run_OLS} \tab logical, if \code{TRUE} the instrument is weak even if it's assumed to be valid. \cr
\tab \cr
\code{run_OLS} \tab logical, if \code{TRUE} the instrument is weak for all tested non empty violation spaces. \cr
\tab \cr
\code{mse_cv} \tab the cross-validation mean squared error of the first stage.
}
}
\description{
Calculates the hat matrix of the first stage.
}
\examples{
n <- 100
Z <- rnorm(n)
X <- rnorm(n)
D <- Z + 2 * Z^2 + X + rnorm(n)
Y <- D + Z + X + rnorm(n)
tsci_boosting(Y = Y, D = D, Z = Z, X = X)
}
