% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsci_forest.R
\name{tsci_forest}
\alias{tsci_forest}
\title{Two Stage Curvature Identification with Random Forests.}
\usage{
tsci_forest(
  Y,
  D,
  Z,
  X,
  intercept = TRUE,
  vio_space = NULL,
  split_prop = 2/3,
  num_trees = NULL,
  mtry = NULL,
  max_depth = NULL,
  min_node_size = NULL,
  forest_save = TRUE,
  str_thol = 10,
  alpha = 0.05,
  nsplits = 1,
  mult_split_method = NULL,
  ncores = 1
)
}
\arguments{
\item{Y}{outcome with dimension n by 1.}

\item{D}{treatment with dimension n by 1.}

\item{Z}{instrument variable with dimension n by 1.}

\item{X}{baseline covariates with dimension n by p.}

\item{intercept}{logic, including the intercept or not in the outcome model, default by TRUE.}

\item{vio_space}{a matrix or a list. If a matrix, then each column corresponds to a violation form of Z; If a list, then each element corresponds to a violation form of Z and must be a matrix of n rows, e.g. (Z^3,Z^2); If NULL, then default by the n by 3 matrix (Z^3, Z^2, Z). Violation form selection will be performed according to provided violation forms, for example, null violation space vs Z vs (Z^2, Z) vs (Z^3, Z^2, Z) in the default case.}

\item{split_prop}{numeric, proportion of observations used to fit the outcome model.}

\item{num_trees}{number of trees in Random Forests, default by 200.}

\item{mtry}{number of covariates to possibly split at in each node of the tree in Random Forests, default by a sequence from round((p+1)/3) to round(2(p+1)/3).}

\item{max_depth}{maximal tree depth in Random Forests, default by 0, which refers to unlimited depth.}

\item{min_node_size}{minimal size of each leaf node in Random Forests, default by the set {5, 10, 15}.}

\item{forest_save}{logical, specifies if the fitted random forest model should be returned.}

\item{str_thol}{minimal value of the threshold of IV strength test, default by 10.}

\item{alpha}{the significance level, default by 0.05.}

\item{nsplits}{numeric, number of times the data will be split.}

\item{mult_split_method}{method to for inference if multi-splitting is performed. Either 'DML' or 'FWER'.}

\item{ncores}{numeric, the number of cores used. \code{mclapply} form the package \code{parallel} will be called. Parallelization is not supported for Windows.}
}
\value{
\describe{
\item{\code{Coef_all}}{a series of point estimators of treatment effect corresponding to different violation spaces and the OLS.}
\item{\code{sd_all}}{standard errors of Coef_all.}
\item{\code{CI_all}}{confidence intervals for the treatment effect corresponding to different violation spaces and the OLS.}
\item{\code{Coef_robust}}{the point estimators corresponding to the violation space selected by the robust comparison.}
\item{\code{sd_robust}}{the standard errors of Coef_robust.}
\item{\code{CI_robust}}{confidence intervals for the treatment effect with the violation space selected by the robust comparison.}
\item{\code{iv_str}}{IV strength corresponding to different violation spaces.}
\item{\code{iv_thol}}{the threshold of IV strength test corresponding to different violation spaces.}
\item{\code{Qmax}}{the index of largest violation space selected by IV strength test. If -1, the IV strength test fails for null violation space and run OLS. If 0, the IV Strength test fails for the null violation space and run TSRF only for null violation space. In other cases, violation space selection is performed.}
\item{\code{q_hat}}{the index of estimated violation space corresponding to Qmax.}
\item{\code{invalidity}}{invalidity of TSLS. If TRUE, the IV is invalid; Otherwise, the IV is valid.}
\item{\code{treatment_model}}{the fitted treatment model. Will only be returned if \code{forest_save} is \code{TRUE} and \code{multi_splitting} is \code{FALSE}.}
}
}
\description{
This function implements Two Stage Curvature Identification with the Random Forests. It tests the IV strength and chooses the best violation form, and also constructs the confidence interval for the treatment effect with the selected violation form.
}
\examples{
\dontrun{
# dimension
p = 10
# sample size
n = 1000
# interaction value
inter.val = 1
# the IV strength
a = 1
# violation strength
tau = 1
f = function(x){a*(1*sin(2*pi*x) + 1.5*cos(2*pi*x))}
rho1=0.5
# function to generate covariance matrix
A1gen=function(rho,p){
  A1=matrix(0,p,p)
  for(i in 1:p){
    for(j in 1:p){
      A1[i,j]=rho^(abs(i-j))
    }
  }
  A1
}
Cov=(A1gen(rho1,p+1))
mu=rep(0,p+1)
# true effect
beta=1
alpha=as.matrix(rep(-0.3,p))
gamma=as.matrix(rep(0.2,p))
inter=as.matrix(c(rep(inter.val,5),rep(0,p-5)))


# generate the data
mu.error=rep(0,2)
Cov.error=matrix(c(1,0.5,0.5,1),2,2)
Error=mvrnorm(n, mu.error, Cov.error)
W.original=mvrnorm(n, mu, Cov)
W=pnorm(W.original)
# instrument variable
Z=W[,1]
# baseline covariates
X=W[,-1]
# generate the treatment variable D
D=f(Z)+X\%*\%alpha+Z*X\%*\%inter+Error[,1]
# generate the outcome variable Y
Y=D*beta+tau*Z+X\%*\%gamma+Error[,2]


# Two Stage Random Forest
output.RF = tsci_forest(Y,D,Z,X)
# point estimates
output.RF$Coef_robust
# standard errors
output.RF$sd_robust
# confidence intervals
output.RF$CI_robust
}
}
